0.0.8:
- Added a HandleManager which behave as an array but with possibility of
removing some of its data at one point and replace it later with new
one, it'll keep track of where a new data must be placed in the array at
any time. Very useful for Data-Oriented prog.
- Added a Reallocation Ratio for HeapArray, SimpleArray & HandleManager,
used when these must be reallocated when full. 1.7x for now.
- Corrected a bug in Simple&HeapArrays reallocation when the
reallocation was based on the count of elems and not the size of the
array.
- Corrected a bug in memory manager when a pointer was deallocated and a
completly new object was inserted into the manager with the exact
(liberated) memory emplacement. This caused memory leaks. Very good
exercise of memory managment in C.