0.0.9:
- Added basic texture handling (with SOIL library).
- Added Texcoords as a basic attrib for mesh. At mesh creation, first
the mesh VBO is filled with position data, then when texcoords data is
given, the data array is reallocated to be filled with pos AND texcoord
data. Then this array is used in vbo. (may need to separate RAM data
array filling and vbo generation in the future to permit dynamic
meshes).
- Fixed bug with reallocation in memory managment (again) when memory
manager was searching for the new allocated pointer in its stack after
reallocation and couldnt find it. now it search for the old pointer.