0.0.10:
- Added a camera system with zoom and pan (for the moment). I plan to
make the zoom target the pos of the mouse and not the center of the
screen.
- Camera now handle the current ProjectionMatrix. and when the window is
resized or the camera moved, the camera warns the device and the
renderer of the changes.
- The device handle the frameTimer now, not the context. The Device now
also handle the WindowResize function
- Modified the EventManager in two ways : (1) the event manager now
store all events that occur in the frame, and send them all in a row at
all listeners at the end of the frame, (2) The listener function has now
a void* ptr as a parameter, so you can pass whatever pointer you want as
a parameter each time the callback is called. See the camera for some
example, or the event.h for doc.
- Modified some code in the matrix operations functions. all works
correctly now and rotation/scale/translation of meshes is flawless.