0.0.11 :
- Added very basic font support(with freetype2). It's very messy
right now and the main contains lot of test code for font rendering.
- Todo : make a font struct containing all that and dome functions to
load and draw text. Futhermore, im using kind of an immediate font
rendering mode right now. I'll have to make Font Atlases to store each
font as a texture (and save it on disk to reload it faster once it has
been loaded).

I had a lot of troubles making the fonts to work with shaders and vbo.
There aren't lot of tutorials or doc and what I found didnt quite work.
Here im loading the rgb(GL_RGB) freetype font to a GL_RED texture and
using this Red channel in the font shader to use it as the alpha channel
of the final color. Then I pass a uniform color to be the text color :).