0.0.12 :
? Text is now drawn on top of everything (with -1.f depth)
? Memory manager arrays size is now fixed in a DEFINE variable (
fixed to 4096 for the moment)
? Enabled Backface Culling. No reason to disable it
? Modified the Copy function of vec2 to be more like the one of mat3.

+ Revamped the Event system for the window resizing callback. Now, the
GLFW Window Resize Callback is in event.c and throw the new window size
to all ResizeListeners of the EventManager (the device and one in main
for the texts for now). This is much more practical.

~ The texts are now correctly displayed when the window size changes.
For now, the current texts are updated from the main. But I plan to add
the text to the Scene and make them all automatically updated everytime
the window size change. (maybe switch them from scene to GUI once I
begin working on that.