0.0.13 :
+ Added the Actor system. Actors are basicly a Game Character, like the
player or any NPC or creature. In it are the required fields defining
them(only their name and entity(rendering) information for now). This
will be extended in the future. This is the really begining of
gameplay. Actors are loaded from a JSON file in data/actors/.
+ Added more documentation for World and Scene.

? World is no more a dynamic structure that can be created and destroyed on the heap. There is now only one World instance, makes sense.
? The ModelMatrices stored in the Entity Array of the scene are now
pointers to matrices, there is no copy anymore. The matrices will be
stored in each Actor I think for now.