0.0.14 :
+ Added the Game Structure/Functions that initialize destroy the game
and all its components (Device, Scene, World), as well as the
Memory manager is in Debug Mode, and the Log.
+ Added a Fullscreen option in the config file.
+ Added a fixed timestep for Physics/IA/Gameplay to make sure the
rendering time does not affect the gameplay. The step is at 100 FPS
(dt = 0.01). The rendering has no step and render as quick as the
machine can.

? The Game_init/destroy functions now init and destroy the Log and
Memory Manager.

~ Fixed a bug where when you add more entities to the scene than the
INITIAL number of entities, some arrays weren't well resized and
provoked segfault. Scene can now admit a lot of entities. I tested
with 5000 on screen entities and get a nice 130 FPS still.