0.1.5 :
+ Remade the Makefile and src folder hierarchy to separate server,
client and common stuff.
+ Net architecture now fully support multi-clients (up to 8 for now)
+ Each client is attributed a sessionID when asking for connection. This
ID is the hash of a random number, to avoid collisions.
+ Net architecture fully handle connection and disconnection.
+ Client and Server are now 2 separate executables. I still need to make
a non-networked version for solo play.
+ ResourceManager is now a static instance initialized by the game. Its
not part of World anymore.
+ The logs now show the time of the msg writing (since client or server
initialization)
+ Memory manager now show the full file name when displaying error,
instead of their first letter only.
+ Added the possibility to have managed data in HandleManager (pointer
data that is allocated and free'd when the handlemanager is destructed
or when asked)

? The architecture is now nearly ready for use with gameplay stuff
transfered over network. When nothing need to be transfered, server and
clients still transmit keepalive(or heartbeats) to maintain the UDP flow
? Client and Server maintain their own log. The Game log has been
removed.
? Actors dont directly load their assets when loaded as resources. This
is done during Game_init. This is done because Actors are a data
structure also present on the server side, as is World.
? You can connect several clients from the same computer. each one will
use a new socket on a new port.